+++ java/src/main/java/com/powsybl/python/commons/PyPowsyblApiHeader.java
@@ -507,7 +507,7 @@ public final class PyPowsyblApiHeader {
         void setAcceptableDuration(int acceptableDuration);
 
         @CField("limit_reduction")
-        float getLimitReduction();
+        double getLimitReduction();
 
         @CField("limit_reduction")
         void setLimitReduction(double limitReduction);
